feat: clean-up code and divide viewer page
