+++ src/ui/features/viewer/components/ViewerEditor.css
@@ -1,0 +1,6 @@
+.msp-plugin {
+  position: relative !important;
+  width: 100% !important;
+  height: 100% !important;
+  border: 1px solid var(--color-grey);
+}
