+++ src/ui/pages/viewer/ViewerPage.css
@@ -1,12 +1,3 @@
-.msp-plugin {
-  position: relative !important;
-  width: 100% !important;
-  height: 100% !important;
-  border: 1px solid var(--color-grey);
-  /* border: 1px solid var(--mantine-primary-color-7);
-  box-shadow: 0 0 20px 1px  var(--mantine-primary-color-2); */
-}
-
 .viewer {
   display: flex;
   flex: 1;
@@ -15,12 +6,4 @@
   height: 100%;
   /* border-style: dashed;
   border-color: rgb(168, 14, 40); */
-}
-
-.viewer-content {
-  position: relative;
-  display: flex;
-  height: 100%;
-  /* border-style: dashed;
-  border-color: rgb(21, 167, 53); */
 }
